# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Authenticates an admin user and sets a session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMe"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Current admin session
      description: Returns the logged-in admin, or 401 if not authenticated
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMe"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin-products]
      summary: List all products (admin)
      description: Returns every product, including inactive ones
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWithDetails"
    post:
      operationId: adminCreateProduct
      tags: [admin-products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithDetails"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin-products]
      summary: Update a product (fields + full variant list)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithDetails"
    delete:
      operationId: adminDeleteProduct
      tags: [admin-products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
  /admin/uploads/presign:
    post:
      operationId: adminPresignUpload
      tags: [admin-products]
      summary: Get a presigned URL to upload a product image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignUploadRequest"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignUploadResponse"
  /admin/products/{id}/images:
    post:
      operationId: adminAddProductImage
      tags: [admin-products]
      summary: Attach an uploaded image to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProductImageRequest"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
  /admin/products/{id}/images/{imageId}:
    delete:
      operationId: adminDeleteProductImage
      tags: [admin-products]
      summary: Remove a product image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
...
  /admin/orders/{id}/ship:
    post:
      operationId: adminShipOrder
      tags: [admin-orders]
      summary: Create a Shiprocket shipment for this order
      description: Requires SHIPROCKET_EMAIL/SHIPROCKET_PASSWORD to be configured on the server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /admin/orders/{id}:
    patch:
      operationId: adminUpdateOrder
      tags: [admin-orders]
      summary: Update order status / tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /admin/packing-recipients:
    get:
      operationId: adminListPackingRecipients
      tags: [admin-packing-recipients]
      summary: List people who get a WhatsApp alert for every new order (packing team)
      responses:
        "200":
          description: Recipients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingRecipientList"
    post:
      operationId: adminCreatePackingRecipient
      tags: [admin-packing-recipients]
      summary: Add a packing-team WhatsApp recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingRecipientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingRecipient"
  /admin/packing-recipients/{id}:
    delete:
      operationId: adminDeletePackingRecipient
      tags: [admin-packing-recipients]
      summary: Remove a packing-team WhatsApp recipient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingRecipientDeleteResult"
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AdminLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AdminMe:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        weightGrams:
          type: integer
        price:
          type: integer
        sku:
          type: string
          nullable: true
        stockQty:
          type: integer
        isDefault:
          type: boolean
      required:
        - id
        - weightGrams
        - price
        - stockQty
        - isDefault
    ProductImage:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
      required:
        - id
        - url
        - sortOrder
        - isPrimary
    ProductWithDetails:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        category:
          type: string
        roast:
          type: string
        notes:
          type: string
        badge:
          type: string
          nullable: true
        active:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
      required:
        - id
        - slug
        - name
        - shortName
        - description
        - category
        - roast
        - notes
        - active
        - variants
        - images
    ProductVariantInput:
      type: object
      properties:
        id:
          type: string
          description: Omit for a new variant
        weightGrams:
          type: integer
        price:
          type: integer
        sku:
          type: string
          nullable: true
        stockQty:
          type: integer
        isDefault:
          type: boolean
      required:
        - weightGrams
        - price
        - stockQty
        - isDefault
    ProductInput:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        category:
          type: string
        roast:
          type: string
        notes:
          type: string
        badge:
          type: string
          nullable: true
        active:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantInput"
      required:
        - slug
        - name
        - shortName
        - description
        - category
        - roast
        - notes
        - active
        - variants
    PresignUploadRequest:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required:
        - filename
        - contentType
    PresignUploadResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        objectKey:
          type: string
        publicUrl:
          type: string
      required:
        - uploadUrl
        - objectKey
        - publicUrl
    AddProductImageRequest:
      type: object
      properties:
        objectKey:
          type: string
        isPrimary:
          type: boolean
      required:
        - objectKey
        - isPrimary
...
    PackingRecipient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - createdAt
    PackingRecipientInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required:
        - name
        - phone
    PackingRecipientList:
      type: array
      items:
        $ref: "#/components/schemas/PackingRecipient"
    PackingRecipientDeleteResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    OrderItemInput:
      type: object
      properties:
        productVariantId:
          type: string
          nullable: true
        name:
          type: string
        price:
          type: integer
        quantity:
          type: integer
      required:
        - name
        - price
        - quantity
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
        notes:
          type: string
          nullable: true
        subtotal:
          type: integer
        shippingMethod:
          type: string
          enum: [shiprocket, india_post]
        shippingFee:
          type: integer
        total:
          type: integer
        paymentMethod:
          type: string
        razorpayOrderId:
          type: string
          description: Present when paid via Razorpay; verified server-side before the order is created
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - phone
        - address
        - city
        - pincode
        - subtotal
        - shippingMethod
        - shippingFee
        - total
        - paymentMethod
        - items
    OrderItemOutput:
      type: object
      properties:
        id:
          type: string
        productVariantId:
          type: string
          nullable: true
        name:
          type: string
        price:
          type: integer
        quantity:
          type: integer
      required:
        - id
        - name
        - price
        - quantity
    OrderWithItems:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
        notes:
          type: string
          nullable: true
        subtotal:
          type: integer
        shippingMethod:
          type: string
        shippingFee:
          type: integer
        total:
          type: integer
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        status:
          type: string
        trackingNumber:
          type: string
          nullable: true
        trackingUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemOutput"
      required:
        - id
        - orderNumber
        - customerName
        - phone
        - address
        - city
        - pincode
        - subtotal
        - shippingMethod
        - shippingFee
        - total
        - paymentMethod
        - paymentStatus
        - status
        - createdAt
        - items
    OrderUpdateInput:
      type: object
      properties:
        status:
          type: string
        trackingNumber:
          type: string
          nullable: true
        trackingUrl:
          type: string
          nullable: true
...